feat(transferencia): auto correct inconsistencies for many animals at the same time

+++ lib/controllers/nascimento_controller.dart
@@ -1,4 +1,5 @@
 import '../models/nascimento.dart';
+import '../session/app_session.dart';
 import '../services/nascimento_service.dart';
 
 class NascimentoController {
@@ -43,6 +44,10 @@ class NascimentoController {
   }
 
   Future<int> delete(int id) {
+    if (!AppSession.isAdmin) {
+      throw StateError('Apenas administradores podem excluir animais.');
+    }
+
     print('[NascimentoController.delete] id=$id');
     return _nascimentoService.delete(id);
   }
@@ -73,6 +78,30 @@ class NascimentoController {
         '[NascimentoController.calcularRestantes] usuarioId=$usuarioId prefixo=$prefixo inicio=$inicio maximo=$maximo');
     return _nascimentoService.calcularRestantes(
       usuarioId: usuarioId,
+      prefixo: prefixo,
+      inicio: inicio,
+      maximo: maximo,
+    );
+  }
+
+  Future<int> obterUltimoNumeroUsadoPorPrefixo({
+    required String prefixo,
+  }) {
+    print(
+        '[NascimentoController.obterUltimoNumeroUsadoPorPrefixo] prefixo="$prefixo"');
+    return _nascimentoService.obterUltimoNumeroUsadoPorPrefixo(
+      prefixo: prefixo,
+    );
+  }
+
+  Future<int> obterPrimeiroNumeroDisponivelPorPrefixo({
+    required String prefixo,
+    required int inicio,
+    required int maximo,
+  }) {
+    print(
+        '[NascimentoController.obterPrimeiroNumeroDisponivelPorPrefixo] prefixo="$prefixo" inicio=$inicio maximo=$maximo');
+    return _nascimentoService.obterPrimeiroNumeroDisponivelPorPrefixo(
       prefixo: prefixo,
       inicio: inicio,
       maximo: maximo,
